delete(application): LottoApplication 삭제
