+++ src/main/java/lotto/application/ScenarioApplication.java
@@ -14,7 +14,6 @@ public class ScenarioApplication {
     private final Calculator calculator;
     private final Printer printer;
     private final PrizeNumber prizeNumber;
-    private final LottoApplication lottoApplication;
 
     public ScenarioApplication(AppConfig appConfig) {
         this.makeNumbersStrategy = appConfig.makeNumbersStrategy();
@@ -22,16 +21,26 @@ public class ScenarioApplication {
         this.calculator = appConfig.calculator();
         this.printer = appConfig.printer();
         this.prizeNumber = appConfig.prizeNumber();
-        this.lottoApplication = appConfig.lottoApplication();
     }
 
     public void run() {
         PurchasePrice purchasePrice = doPurchase();
         LottoQuantity lottoQuantity = LottoQuantity.findQuantity(purchasePrice.value());
-        Lottos lottos = lottoApplication.buyLottos(lottoQuantity);
-        WinNumbers winNumbers = lottoApplication.readAllWinNumbers();
+        Lottos lottos = buyLottos(lottoQuantity);
+        WinNumbers winNumbers = readAllWinNumbers();
         prizeNumber.findWinningLottos(winNumbers, lottos);
         calculateResult(purchasePrice, prizeNumber);
+    }
+
+    public Lottos buyLottos(LottoQuantity lottoQuantity) {
+        Lottos lottos = Lottos.from(lottoQuantity, makeNumbersStrategy);
+        printer.printPurchaseResult(lottoQuantity.value(), lottos);
+        return lottos;
+    }
+
+    public WinNumbers readAllWinNumbers() {
+        WinNumbers winNumbersWithoutBonusNumber = readWinNumbers();
+        return readBonusNumber(winNumbersWithoutBonusNumber);
     }
 
     private PurchasePrice doPurchase() {
@@ -51,5 +60,27 @@ public class ScenarioApplication {
         int totalPrize = prizeNumber.calculateTotalPrize();
         double profit = calculator.calculateProfit(totalPrize, purchasePrice.value());
         printer.printPrizeResult(prizeNumber.sortByRank(), profit);
+    }
+
+    private WinNumbers readWinNumbers() {
+        printer.print("\n당첨 번호를 입력해 주세요.");
+        String originWinNumbers = reader.read();
+        try {
+            return WinNumbers.winNumbersFrom(originWinNumbers);
+        } catch (IllegalArgumentException e) {
+            printer.print(e.getMessage());
+            return readWinNumbers();
+        }
+    }
+
+    private WinNumbers readBonusNumber(WinNumbers winNumbersWithOutBonusNumber) {
+        printer.print("\n보너스 번호를 입력해 주세요.");
+        String bonusNumber = reader.read();
+        try {
+            return winNumbersWithOutBonusNumber.bonusNumberFrom(bonusNumber);
+        } catch (IllegalArgumentException e) {
+            printer.print(e.getMessage());
+            return readBonusNumber(winNumbersWithOutBonusNumber);
+        }
     }
 }
